[refactor] enum 클래스와 class 클래스 리팩토링 및 임포트
