+++ src/api/dto/response/ProblemListResponseDto.ts
@@ -1,5 +1,5 @@
 import { ProblemPage } from "../../../common/class";
-import { RESP_STATUS } from "../../../common/constants";
+import { RESP_STATUS } from "../../../common/enums/response-status.enum";
 
 class ProblemListResponseDto {
   validCheck;
