+++ src/common/enums/icon.enum.ts
@@ -1,0 +1,5 @@
+export enum IconType {
+  EMOJI = "EMOJI",
+  EXTERNAL = "EXTERNAL",
+  FILE = "FILE",
+}
