+++ src/common/enums/response-status.enum.ts
@@ -1,0 +1,7 @@
+export enum RESP_STATUS {
+  SUCCESS = "success",
+  FAILED = "failed",
+  NOT_FOUND = "not_found",
+  INVALID = "invalid",
+  UNAUTHORIZED = "unauthorized",
+}
