+++ src/common/enums/site.enum.ts
@@ -1,0 +1,11 @@
+export enum SiteHost {
+  BOJ = "www.acmicpc.net",
+  PROGRAMMERS = "school.programmers.co.kr",
+}
+
+export enum SiteType {
+  BOJ = "BOJ",
+  PROGRAMMERS = "PROGRAMMERS",
+  PROGRAMMERS_SQL = "PROGRAMMERS_SQL",
+  SWEA = "SWEA",
+}
