+++ src/common/enums/storage.enum.ts
@@ -1,0 +1,8 @@
+export enum StorageKey {
+  NOTION_API_KEY = "notion_api_key",
+  DATABASE_ID = "database_id",
+  PROBLEM_PAGE_LIST = "problem_page_list",
+  RESP_STATUS = "response_status",
+  PROBLEM_OPTIONS = "problem_options",
+  FILTERED_PROBLEM = "filtered_problem",
+}
