+++ src/common/models/user-status.model.ts
@@ -1,0 +1,10 @@
+import { RESP_STATUS } from "../enums/response-status.enum";
+
+export interface IUserStatus {
+  isNotionLinked: boolean;
+  respStatus: RESP_STATUS;
+}
+
+export class UserStatus implements IUserStatus {
+  constructor(public isNotionLinked: boolean, public respStatus: RESP_STATUS) {}
+}
