+++ src/popup/tabs/TabList.tsx
@@ -3,7 +3,7 @@ import { Tab, Tabs } from "react-bootstrap";
 import ProblemInsertTab from "./ProblemInsertTab";
 import RandomSelectTab from "./RandomSelectTab";
 import "../popup.css";
-import { ProblemPage } from "../../common/class";
+import { IProblemPage } from "../../common/models/problem-page.model";
 
 enum TabKey {
   CURRENT_PROBLEM = "currentProblem",
@@ -12,7 +12,7 @@ enum TabKey {
 
 interface ProblemPageInfo {
   isExist: boolean;
-  problemPage: ProblemPage;
+  problemPage: IProblemPage;
 }
 
 interface Props {
